fix: optimize shortcuts
